refactor(PlayerStatus): 상금 계산 로직 할당
