+++ src/main/java/blackjack/domain/participant/playerstatus/Blackjack.java
@@ -1,11 +1,10 @@
 package blackjack.domain.participant.playerstatus;
 
-import blackjack.domain.prizecalculator.BlackjackCalculator;
-import blackjack.domain.prizecalculator.PrizeCalculator;
-
 public final class Blackjack extends CalculableStatus {
 
     private static final Blackjack INSTANCE = new Blackjack();
+    private static final double WIN_EARNING_RATE = 1.5;
+    private static final int PUSH_EARNING_RATE = 0;
 
     private Blackjack() {
     }
@@ -15,7 +14,11 @@ public final class Blackjack extends CalculableStatus {
     }
 
     @Override
-    public PrizeCalculator findCalculator() {
-        return BlackjackCalculator.getInstance();
+    public double calculateProfit(final int playerScore, final int dealerScore, final boolean dealerBlackjack,
+                                  final double bettingAmount) {
+        if (dealerBlackjack) {
+            return bettingAmount * PUSH_EARNING_RATE;
+        }
+        return bettingAmount * WIN_EARNING_RATE;
     }
 }
