--- src/main/java/blackjack/domain/prizecalculator/PrizeCalculator.java
@@ -1,7 +1,0 @@
-package blackjack.domain.prizecalculator;
-
-public interface PrizeCalculator {
-
-    double calculate(final int playerScore, final int dealerScore, final boolean dealerBlackjack,
-                     final double bettingAmount);
-}
