+++ src/test/java/blackjack/domain/participant/playerstatus/BustTest.java
@@ -1,13 +1,12 @@
-package blackjack.domain.prizecalculator;
+package blackjack.domain.participant.playerstatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.participant.playerstatus.Bust;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class BustCalculatorTest {
+class BustTest {
 
     @ParameterizedTest
     @DisplayName("플레이어가 BUST일 때 상금을 계산한다.")
@@ -15,12 +14,13 @@ class BustCalculatorTest {
             "22:false:400:-400"}, delimiter = ':')
     void calculatePrize_bust(int dealerScore, boolean dealerBlackjack, int bettingAmount, int expected) {
         // give
-        final PrizeCalculator prizeCalculator = Bust.getInstance().findCalculator();
+        final CalculableStatus status = Bust.getInstance();
 
         // when
-        final double actual = prizeCalculator.calculate(22, dealerScore, dealerBlackjack, bettingAmount);
+        final double actual = status.calculateProfit(22, dealerScore, dealerBlackjack, bettingAmount);
 
         // then
         assertThat(actual).isEqualTo(expected);
     }
+
 }
